requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.3
black==24.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dill==0.3.8
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.109.2
h11==0.14.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.16
starlette==0.36.3
tomlkit==0.12.3
typing_extensions==4.9.0
uvicorn==0.27.1
